palworld/requirements.txt
python-dotenv==1.1.1
requests==2.32.4
discord==2.3.2
